+++ src/hotspot/share/compiler/compileBroker.cpp
@@ -1061,7 +1061,7 @@ void CompileBroker::possibly_add_compiler_threads(JavaThread* THREAD) {
   if (new_c2_count <= old_c2_count && new_c1_count <= old_c1_count) return;
 
   // Now, we do the more expensive operations.
-  size_t free_memory = 0;
+  uint64_t free_memory = 0;
   // Return value ignored - defaulting to 0 on failure.
   (void)os::free_memory(free_memory);
   // If SegmentedCodeCache is off, both values refer to the single heap (with type CodeBlobType::All).
